move index.ts to bot.ts
